:fire: removed old controller
