+++ packages/youtube_player_iframe/example/lib/main.dart
@@ -6,7 +6,6 @@ import 'dart:developer';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:youtube_player_iframe/youtube_player_iframe.dart';
 
 import 'widgets/meta_data_section.dart';
@@ -50,37 +49,38 @@ class _YoutubeAppDemoState extends State<YoutubeAppDemo> {
   void initState() {
     super.initState();
     _controller = YoutubePlayerController(
-      initialVideoId: 'tcodrIK2P_I',
-      params: const YoutubePlayerParams(
-        playlist: [
-          'nPt8bK2gbaU',
-          'K18cpp_-gP8',
-          'iLnmTe5Q2Qw',
-          '_WoCV4c6XOE',
-          'KmzdUe0RSJo',
-          '6jZDSSZZxjQ',
-          'p2lYr3vM_1w',
-          '7QUtEmBT_-w',
-          '34_PXCzGw1M',
-        ],
-        startAt: const Duration(minutes: 1, seconds: 36),
-        showControls: true,
-        showFullscreenButton: true,
-        desktopMode: false,
-        privacyEnhanced: true,
-        useHybridComposition: true,
-      ),
-    );
-    _controller.onEnterFullscreen = () {
-      SystemChrome.setPreferredOrientations([
-        DeviceOrientation.landscapeLeft,
-        DeviceOrientation.landscapeRight,
-      ]);
-      log('Entered Fullscreen');
-    };
-    _controller.onExitFullscreen = () {
-      log('Exited Fullscreen');
-    };
+        // initialVideoId: 'tcodrIK2P_I',
+        // params: const YoutubePlayerParams(
+        //   /*   playlist: [
+        //     'nPt8bK2gbaU',
+        //     'K18cpp_-gP8',
+        //     'iLnmTe5Q2Qw',
+        //     '_WoCV4c6XOE',
+        //     'KmzdUe0RSJo',
+        //     '6jZDSSZZxjQ',
+        //     'p2lYr3vM_1w',
+        //     '7QUtEmBT_-w',
+        //     '34_PXCzGw1M',
+        //   ],*/
+        //   startAt: const Duration(minutes: 1, seconds: 36),
+        //   showControls: true,
+        //   showFullscreenButton: true,
+        //   desktopMode: false,
+        //   privacyEnhanced: true,
+        //   useHybridComposition: true,
+        // ),
+        );
+    _controller.loadVideoById(videoId: 'tcodrIK2P_I');
+    // _controller.onEnterFullscreen = () {
+    //   SystemChrome.setPreferredOrientations([
+    //     DeviceOrientation.landscapeLeft,
+    //     DeviceOrientation.landscapeRight,
+    //   ]);
+    //   log('Entered Fullscreen');
+    // };
+    // _controller.onExitFullscreen = () {
+    //   log('Exited Fullscreen');
+    // };
   }
 
   @override
@@ -111,7 +111,7 @@ class _YoutubeAppDemoState extends State<YoutubeAppDemo> {
                 Stack(
                   children: [
                     player,
-                    Positioned.fill(
+                    /*Positioned.fill(
                       child: YoutubeValueBuilder(
                         controller: _controller,
                         builder: (context, value) {
@@ -143,7 +143,7 @@ class _YoutubeAppDemoState extends State<YoutubeAppDemo> {
                           );
                         },
                       ),
-                    ),
+                    ),*/
                   ].take(1).toList(),
                 ),
                 const Controls(),
@@ -157,7 +157,7 @@ class _YoutubeAppDemoState extends State<YoutubeAppDemo> {
 
   @override
   void dispose() {
-    _controller.close();
+    //_controller.close();
     super.dispose();
   }
 }
