+++ packages/youtube_player_iframe/example/lib/widgets/meta_data_section.dart
@@ -53,7 +53,7 @@ class MetaDataSection extends StatelessWidget {
                           .toList(),
                       onChanged: (double? newValue) {
                         if (newValue != null) {
-                          context.ytController.setPlaybackRate(newValue);
+                          //context.ytController.setPlaybackRate(newValue);
                         }
                       },
                     );
