+++ packages/youtube_player_iframe/lib/src/controller/youtube_player_controller.dart
@@ -3,8 +3,9 @@ import 'dart:convert';
 
 import 'package:flutter/services.dart';
 import 'package:webview_flutter/webview_flutter.dart';
+import 'package:youtube_player_iframe/youtube_player_iframe.dart';
 
-import '../iframe_api/youtube_player_iframe_api.dart';
+import '../player_value.dart';
 import '../web_registrar/register_web_webview_stub.dart'
     if (dart.library.html) '../web_registrar/register_web_webview.dart';
 import 'youtube_player_event_handler.dart';
@@ -20,6 +21,13 @@ class YoutubePlayerController implements YoutubePlayerIFrameAPI {
 
   late final YoutubePlayerEventHandler _eventHandler;
   late final Set<JavascriptChannel> javaScriptChannels;
+
+  final StreamController<YoutubePlayerValue> _valueController =
+      StreamController.broadcast();
+  YoutubePlayerValue _value = YoutubePlayerValue();
+
+  /// The [YoutubePlayerValue].
+  YoutubePlayerValue get value => _value;
 
   Future<WebViewController> get webViewController {
     return _webViewControllerCompleter.future;
@@ -125,17 +133,22 @@ class YoutubePlayerController implements YoutubePlayerIFrameAPI {
     );
   }
 
-  Future<void> load(
-    WebViewController controller, {
+  Future<void> init(WebViewController controller) async {
+    _webViewControllerCompleter.complete(controller);
+    await load(params: const YoutubePlayerParams());
+  }
+
+  Future<void> load({
+    required YoutubePlayerParams params,
     String? baseUrl,
   }) async {
-    _webViewControllerCompleter.complete(controller);
     final playerHtml = await rootBundle.loadString(
       'packages/youtube_player_iframe/assets/player.html',
     );
 
+    final controller = await _webViewControllerCompleter.future;
     await controller.loadHtmlString(
-      playerHtml.replaceFirst('<<playerVars>>', "{'playsinline': 1}"),
+      playerHtml.replaceFirst('<<playerVars>>', params.toJson()),
       baseUrl: baseUrl,
     );
   }
@@ -150,5 +163,59 @@ class YoutubePlayerController implements YoutubePlayerIFrameAPI {
     final varArgs = data == null ? '' : jsonEncode(data);
 
     return controller.runJavascript('player.$functionName($varArgs);');
+  }
+
+  /// MetaData for the currently loaded or cued video.
+  YoutubeMetaData get metadata => _value.metaData;
+
+  /// Listen to updates in [YoutubePlayerController].
+  StreamSubscription<YoutubePlayerValue> listen(
+    void Function(YoutubePlayerValue event)? onData, {
+    Function? onError,
+    void Function()? onDone,
+    bool? cancelOnError,
+  }) {
+    return _valueController.stream.listen(
+      (value) {
+        _value = value;
+        onData?.call(value);
+      },
+      onError: onError,
+      onDone: onDone,
+      cancelOnError: cancelOnError,
+    );
+  }
+
+  /// Converts fully qualified YouTube Url to video id.
+  ///
+  /// If videoId is passed as url then no conversion is done.
+  static String? convertUrlToId(String url, {bool trimWhitespaces = true}) {
+    if (!url.contains("http") && (url.length == 11)) return url;
+    if (trimWhitespaces) url = url.trim();
+
+    for (var regex in [
+      r'^https:\/\/(?:www\.|m\.)?youtube\.com\/watch\?v=([_\-a-zA-Z0-9]{11}).*$',
+      r'^https:\/\/(?:www\.|m\.)?youtube(?:-nocookie)?\.com\/embed\/([_\-a-zA-Z0-9]{11}).*$',
+      r'^https:\/\/youtu\.be\/([_\-a-zA-Z0-9]{11}).*$',
+    ]) {
+      Match? match = RegExp(regex).firstMatch(url);
+      if (match != null && match.groupCount >= 1) return match.group(1);
+    }
+
+    return null;
+  }
+
+  /// Grabs YouTube video's thumbnail for provided video id.
+  ///
+  /// If [webp] is true, webp version of the thumbnail will be retrieved,
+  /// Otherwise a JPG thumbnail.
+  static String getThumbnail({
+    required String videoId,
+    String quality = ThumbnailQuality.standard,
+    bool webp = true,
+  }) {
+    return webp
+        ? 'https://i3.ytimg.com/vi_webp/$videoId/$quality.webp'
+        : 'https://i3.ytimg.com/vi/$videoId/$quality.jpg';
   }
 }
